+++ lib/rest/bulkexports/v1/export/exportCustomJob.d.ts
@@ -214,7 +214,9 @@ interface ExportCustomJobResource {
   details: object;
   email: string;
   end_day: string;
+  estimated_completion_time: string;
   friendly_name: string;
+  job_queue_position: string;
   job_sid: string;
   resource_type: string;
   start_day: string;
@@ -243,7 +245,9 @@ declare class ExportCustomJobInstance extends SerializableClass {
   details: any;
   email: string;
   endDay: string;
+  estimatedCompletionTime: string;
   friendlyName: string;
+  jobQueuePosition: string;
   jobSid: string;
   resourceType: string;
   startDay: string;