+++ lib/rest/supersim/v1/fleet.d.ts
@@ -28,12 +28,16 @@ declare function FleetList(version: V1): FleetListInstance;
  * @property commandsMethod - A string representing the HTTP method to use when making a request to `commands_url`
  * @property commandsUrl - The URL that will receive a webhook when a Super SIM in the Fleet is used to send an SMS from your device to the Commands number
  * @property networkAccessProfile - The SID or unique name of the Network Access Profile of the Fleet
+ * @property smsCommandsMethod - A string representing the HTTP method to use when making a request to `sms_commands_url`
+ * @property smsCommandsUrl - The URL that will receive a webhook when a Super SIM in the Fleet is used to send an SMS from your device to the SMS Commands number
  * @property uniqueName - An application-defined string that uniquely identifies the resource
  */
 interface FleetInstanceUpdateOptions {
   commandsMethod?: string;
   commandsUrl?: string;
   networkAccessProfile?: string;
+  smsCommandsMethod?: string;
+  smsCommandsUrl?: string;
   uniqueName?: string;
 }
 
@@ -166,6 +170,9 @@ interface FleetListInstance {
  * @property dataEnabled - Defines whether SIMs in the Fleet are capable of using data connectivity
  * @property dataLimit - The total data usage (download and upload combined) in Megabytes that each Sim resource assigned to the Fleet resource can consume
  * @property networkAccessProfile - The SID or unique name of the Network Access Profile of the Fleet
+ * @property smsCommandsEnabled - Defines whether SIMs in the Fleet are capable of sending and receiving machine-to-machine SMS via Commands
+ * @property smsCommandsMethod - A string representing the HTTP method to use when making a request to `sms_commands_url`
+ * @property smsCommandsUrl - The URL that will receive a webhook when a Super SIM in the Fleet is used to send an SMS from your device to the SMS Commands number
  * @property uniqueName - An application-defined string that uniquely identifies the resource
  */
 interface FleetListInstanceCreateOptions {
@@ -175,6 +182,9 @@ interface FleetListInstanceCreateOptions {
   dataEnabled?: boolean;
   dataLimit?: number;
   networkAccessProfile: string;
+  smsCommandsEnabled?: boolean;
+  smsCommandsMethod?: string;
+  smsCommandsUrl?: string;
   uniqueName?: string;
 }
 
@@ -256,6 +266,9 @@ interface FleetResource {
   date_updated: Date;
   network_access_profile_sid: string;
   sid: string;
+  sms_commands_enabled: boolean;
+  sms_commands_method: string;
+  sms_commands_url: string;
   unique_name: string;
   url: string;
 }
@@ -333,6 +346,9 @@ declare class FleetInstance extends SerializableClass {
   fetch(callback?: (error: Error | null, items: FleetInstance) => any): Promise<FleetInstance>;
   networkAccessProfileSid: string;
   sid: string;
+  smsCommandsEnabled: boolean;
+  smsCommandsMethod: string;
+  smsCommandsUrl: string;
   /**
    * Provide a user-friendly representation
    */