+++ spec/integration/rest/bulkexports/v1/export/job.spec.js
@@ -63,7 +63,9 @@ describe('Job', function() {
           'webhook_url': 'webhook_url',
           'webhook_method': 'webhook_method',
           'email': 'email',
-          'resource_type': 'resource_type'
+          'resource_type': 'resource_type',
+          'job_queue_position': '1',
+          'estimated_completion_time': '2021-03-15T20:20:14.547'
       };
 
       holodeck.mock(new Response(200, body));