+++ spec/integration/rest/taskrouter/v1/workspace/task.spec.js
@@ -99,8 +99,9 @@ describe('Task', function() {
     function(done) {
       holodeck.mock(new Response(500, {}));
 
+      var opts = {ifMatch: 'if_match'};
       var promise = client.taskrouter.v1.workspaces('WSXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
-                                        .tasks('WTXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX').update();
+                                        .tasks('WTXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX').update(opts);
       promise.then(function() {
         throw new Error('failed');
       }, function(error) {
@@ -112,9 +113,11 @@ describe('Task', function() {
       var sid = 'WTXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX';
       var url = `https://taskrouter.twilio.com/v1/Workspaces/${workspaceSid}/Tasks/${sid}`;
 
+      var headers = {'If-Match': 'if_match'};
       holodeck.assertHasRequest(new Request({
         method: 'POST',
-        url: url
+        url: url,
+        headers: headers
       }));
     }
   );